plugins/lugh/skills/new-ds-project/template/pyproject.toml
[project]
name = "__NAME__"
version = "0.1.0"
description = "__TITLE__ — data science project"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "firstname.lastname", email = "firstname.lastname@example.com" }]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "dvc>=3.30.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pre-commit>=3.6.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

viz = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

notebook = [
    "marimo>=0.9.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[project.scripts]
__NAME__ = "__PACKAGE__.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
